Tehtävä 4.21

+++ controllers/blogs.js
@@ -52,8 +52,20 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.BLOG_SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === user.id) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    }
+    else {
+      return response.status(400).json({ error: 'Blog can be deleted only by user, who creted it' })
+    }
   } catch (exception) {
     console.log(exception)
     response.status(500).json({ error: 'Unexpected error' })
